test_app/drf_login/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-environ
whitenoise
uvicorn